fix(tiptap): add " (nouvelle fenêtre)" to links
- as they all open in a new window
- uses a `<span class="fr-sr-only">`

Note:
There is a Prosemirror limitation to how marks (a link is a mark) are rendered: "Content
hole must be the only child of its parent node". See That’s why we must have
a `<span>` element wrapping the text content of the link (set inside the
"hole", which is represented by a "0").